pyproject.toml
[tool.poetry]
name = "gitwp"
version = "0.1.0"
description = ""
authors = ["Dave Gaeddert <dave.gaeddert@dropseed.dev>"]
readme = "README.md"

[tool.poetry.scripts]
gitwp = "gitwp.cli:cli"

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.3"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.263"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
